requirements.txt
opencv-python>=3.4
python-crontab>=2.2.8
python-configparser>=3.3.0
